Closes #16 Migrate database IDs to u64
Update entity models and database logic to use `u64` for Discord IDs,
aligning with Serenity's `UserId` type.

Added stricter clippy lints and fixed caught lints

+++ src/entities/items.rs
@@ -6,13 +6,13 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "items")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub description: String,
     pub item_type: String,
     pub quantity: i64,
     pub price: i64,
-    pub origin_cid: Option<i64>,
+    pub origin_cid: Option<u64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
