+++ src/users/inventory/list.rs
@@ -6,7 +6,7 @@ use poise::serenity_prelude::CreateEmbed;
 pub async fn inventory(ctx: poise::Context<'_, crate::Data, Error>) -> Result<(), Error> {
     let author = ctx.author();
     let items =
-        crate::users::inventory::db::db_get_inventory(author.id.get() as i64, &ctx.data().database)
+        crate::users::inventory::db::db_get_inventory(author.id.get(), &ctx.data().database)
             .await?;
 
     if items.is_empty() {
